Module Tar_mirage

Tar for Mirage

module Make_KV_RO (BLOCK : Mirage_block.S) : sig ... end

Construct a read-only key-value store from an existing block device containing tar-format data.

module Make_KV_RW (BLOCK : Mirage_block.S) : sig ... end

Construct a read-write key-value store from an existing block device containing tar-format data. Note that it is append-only meaning removing or renaming files is currently unsupported and will return an error.